Makes the long bone ratio separate from assessment.
Since the long bone ratio is a combination of the long bone and its
assessment, the property bound to the radio buttons became the
"assessment", and a new property was made for the ratio. This decreases
the amount of property changes being raised, reduces some code, and
better reflects the interplay between the properties.